requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.0.0
supabase==2.11.0
python-docx==1.1.2
fpdf2==2.8.2
openpyxl==3.1.5
pandas==2.2.3
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
aiosmtplib==3.0.2
jinja2==3.1.5
